[OPE-46] Keep RHIME preparation focused on science
