+++ CHANGELOG.md
@@ -8,8 +8,8 @@
   PyMC materialization into a cohesive preparation module while keeping their
   scientific order visible in the runners. `run_rhime` and
   `run_rhime_multisector` can now resume from borrowed external
-  `RhimeMergedData` without OpenGHG or cache I/O, and durable prepared-input
-  artifacts retain model-independent JSON preparation provenance.
+  `RhimeMergedData` without OpenGHG or cache I/O while preserving the existing
+  prepared-input serialization contract.
   [OPE-46](https://linear.app/openghg-inversions/issue/OPE-46/p0-w4-split-rhime-preparation-into-named-scientific-stages)
 
 - Added a package-shaped RHIME consumer example for projects generated by the
